Moved 100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-long long i, n;
+long int i, n;
 
 n = 612852475143;
 for (i = 2; i < n; i++)
